--- src/fs/sample.ts
@@ -1,12 +1,0 @@
-import * as FileSystem from "../fs/types.js"
-import { Branch } from "../path/index.js"
-
-
-export async function addSampleData(fs: FileSystem.API): Promise<void> {
-  await fs.mkdir({ directory: [ Branch.Private, "Apps" ] })
-  await fs.mkdir({ directory: [ Branch.Private, "Audio" ] })
-  await fs.mkdir({ directory: [ Branch.Private, "Documents" ] })
-  await fs.mkdir({ directory: [ Branch.Private, "Photos" ] })
-  await fs.mkdir({ directory: [ Branch.Private, "Video" ] })
-  await fs.publish()
-}
